requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
decorator==5.2.1
Deprecated==1.3.1
distro==1.9.0
executing==2.2.1
finmind==1.9.3
Flask==3.1.2
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.7
langchain-core==1.2.7
langchain-deepseek==1.0.1
langchain-openai==1.1.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
line-bot-sdk==3.22.0
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
multidict==6.7.1
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.4.1
openai==2.16.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pandas==3.0.0
parso==0.8.5
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyecharts==2.0.9
Pygments==2.19.2
pyngrok==7.5.0
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.10.2
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
ta==0.5.25
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
wcwidth==0.5.0
websockets==16.0
Werkzeug==3.1.5
win32_setctime==1.2.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
yfinance==1.1.0
zipp==3.23.0
zstandard==0.25.0
